+++ src/main/java/leets/leenk/domain/user/application/usecase/UserSettingUsecase.java
@@ -6,6 +6,9 @@ import leets.leenk.domain.user.application.mapper.UserSettingMapper;
 import leets.leenk.domain.user.domain.entity.User;
 import leets.leenk.domain.user.domain.entity.UserSetting;
 import leets.leenk.domain.user.domain.service.*;
+import leets.leenk.domain.user.domain.service.user.UserGetService;
+import leets.leenk.domain.user.domain.service.usersetting.UserSettingGetService;
+import leets.leenk.domain.user.domain.service.usersetting.UserSettingUpdateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
