--- src/main/java/leets/leenk/domain/user/domain/service/UserDeleteService.java
@@ -1,12 +1,0 @@
-package leets.leenk.domain.user.domain.service;
-
-import leets.leenk.domain.user.domain.entity.User;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDeleteService {
-
-    public void delete(User user) {
-        user.delete();
-    }
-}
