+++ src/main/java/leets/leenk/domain/user/domain/service/user/UserGetService.java
@@ -1,4 +1,4 @@
-package leets.leenk.domain.user.domain.service;
+package leets.leenk.domain.user.domain.service.user;
 
 import leets.leenk.domain.user.application.exception.UserNotFoundException;
 import leets.leenk.domain.user.domain.entity.User;
@@ -18,7 +18,7 @@ public class UserGetService {
     private final UserRepository userRepository;
 
     public User findById(long userId) {
-        return userRepository.findById(userId)
+        return userRepository.findByIdAndDeleteDateIsNull(userId)
                 .orElseThrow(UserNotFoundException::new);
     }
 
@@ -27,14 +27,14 @@ public class UserGetService {
     }
 
     public List<User> findAll(List<Long> userIds) {
-        return userRepository.findAllByIdIn(userIds);
+        return userRepository.findAllByIdInAndDeleteDateIsNull(userIds);
     }
 
     public List<User> findAll() {
-        return userRepository.findAllByOrderByName();
+        return userRepository.findAllByDeleteDateIsNullOrderByName();
     }
 
     public Slice<User> findAll(Pageable pageable) {
-        return userRepository.findAllByOrderByName(pageable);
+        return userRepository.findAllByDeleteDateIsNullOrderByName(pageable);
     }
 }
