+++ src/main/java/leets/leenk/domain/user/domain/service/user/UserSaveService.java
@@ -1,4 +1,4 @@
-package leets.leenk.domain.user.domain.service;
+package leets.leenk.domain.user.domain.service.user;
 
 import leets.leenk.domain.user.domain.entity.User;
 import leets.leenk.domain.user.domain.repository.UserRepository;
